Update app structure and styles, add Recommendation component
The commit introduces a more structured layout by wrapping main components inside the ThemeProvider and MainContainer, and adds the Recommendation section. In addition to this, it cleans up the import statements and adjusts the global styles to enhance the overall look and feel. It also incorporates new components such as Recommendation and Education, hence adding more functionality and features tyeso the app.